packages refactoring of the MatchData, move regexp in smarts
