+++ src/main/java/com/antkorwin/springtestmongo/internal/expect/match/simple/MatchDataFactory.java
@@ -1,5 +1,6 @@
-package com.antkorwin.springtestmongo.internal.expect.match;
+package com.antkorwin.springtestmongo.internal.expect.match.simple;
 
+import com.antkorwin.springtestmongo.internal.expect.match.MatchData;
 import com.fasterxml.jackson.databind.node.JsonNodeType;
 
 /**
