+++ src/test/java/com/antkorwin/springtestmongo/internal/expect/match/simple/MatchDataFactoryTest.java
@@ -1,5 +1,7 @@
-package com.antkorwin.springtestmongo.internal.expect.match;
+package com.antkorwin.springtestmongo.internal.expect.match.simple;
 
+import com.antkorwin.springtestmongo.internal.expect.match.MatchData;
+import com.antkorwin.springtestmongo.internal.expect.match.simple.*;
 import com.fasterxml.jackson.databind.node.JsonNodeType;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
