+++ src/test/java/com/antkorwin/springtestmongo/internal/expect/match/simple/MatchNumberTest.java
@@ -1,5 +1,6 @@
-package com.antkorwin.springtestmongo.internal.expect.match;
+package com.antkorwin.springtestmongo.internal.expect.match.simple;
 
+import com.antkorwin.springtestmongo.internal.expect.match.simple.MatchNumber;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
