+++ src/test/java/com/antkorwin/springtestmongo/internal/expect/match/simple/MatchObjectsTest.java
@@ -1,7 +1,8 @@
-package com.antkorwin.springtestmongo.internal.expect.match;
+package com.antkorwin.springtestmongo.internal.expect.match.simple;
 
 import java.util.stream.Stream;
 
+import com.antkorwin.springtestmongo.internal.expect.match.simple.MatchObjects;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
